requirements.txt
faster-whisper
yt-dlp
moviepy
opencv-python
mediapipe==0.10.14
flet==0.23.2
python-dotenv
colorama
requests
psutil